Initial commit: Flask User Management API with JWT authentication

+++ services.py
@@ -1,4 +1,5 @@
 from models import User
+from werkzeug.security import generate_password_hash
 from extensions import db
 
 
@@ -11,15 +12,27 @@ def get_user(user_id):
     return User.query.get(user_id)
 
 
+def get_user_by_email(email):
+    return User.query.filter_by(email=email).first()
 
-def create_user(name, email):
-    new_user = User(name=name, email=email)
+## Using hashed_password 
+
+def create_user(name, email,password):
+    hashed_password = generate_password_hash(password)
+    new_user = User(
+        name=name, 
+        email=email,
+        password=hashed_password
+        )
+    
     db.session.add(new_user)
     db.session.commit()
     return new_user
 
 def update_user(user_id,name,email):
     user = User.query.get(user_id)
+    if not user:
+        return None
     user.name = name
     user.email = email
     db.session.commit()
@@ -27,12 +40,12 @@ def update_user(user_id,name,email):
 
 def delete_user(user_id):
     user = User.query.get(user_id)
+    if not user:
+        return False
     db.session.delete(user)
     db.session.commit()
-    return user
+    return True
 
 
-def get_user_by_email(email):
-    user =  User.query.filter_by(email=email).first()
-    return user
+
